Ok, now, fix the module shit

+++ modules/gy-271/core/test.py
@@ -5,7 +5,7 @@ import os
 f = open(os.path.dirname(os.path.realpath(__file__)) + '/../internals/test_data.json', 'r')
 response = json.loads(f.read())
 
-if isinstance(response['x'], float) and isinstance(response['y'], float) and isinstance(response['z'], float) :
+if isinstance(response['x'], int) and isinstance(response['y'], int) and isinstance(response['z'], int) :
 	print "Looks like it's working well. Good news!"
 else:
 	print "Looks like it's is NOT working. Bad news!"
